20230120更新：修改uid的类型为qlonglong，解决了过长uid用户数据读取错误的问题

+++ modmenu.h
@@ -33,7 +33,7 @@ public:
     explicit MODmenu(QWidget *parent = nullptr);
     ~MODmenu();
     void flashtablefromvector();
-    int uidFindLocation(int uid);
+    int uidFindLocation(qlonglong uid);
     void savejson();
     void checkfirstrun();
     helpmenu *help;
